pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tilapia-freshness-evaluator"
version = "1.0.0"
description = "A computer vision system for evaluating tilapia fish freshness"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "John Elway Cortez", email = "elwaycortez@gmail.com"},
    {name = "Dustin Uriel Obispo"},
    {name = "King Red Sanchez"},
    {name = "Briel Aldous Viola"}
]
maintainers = [
    {name = "John Elway Cortez", email = "elwaycortez@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["computer-vision", "fish-freshness", "yolo", "opencv", "image-processing"]
requires-python = ">=3.11"
dependencies = [
    "opencv-python==4.12.0.88",
    "pillow==12.0.0",
    "numpy==2.2.6",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "pre-commit>=2.17.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/Evrouin/tilapia-fish-freshness-evaluation-system"
Repository = "https://github.com/Evrouin/tilapia-fish-freshness-evaluation-system"
Issues = "https://github.com/Evrouin/tilapia-fish-freshness-evaluation-system/issues"

[project.scripts]
tilapia-freshness = "tilapia_freshness.gui.main_window:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
tilapia_freshness = ["assets/*"]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
disable_error_code = ["index", "unused-ignore"]

[[tool.mypy.overrides]]
module = "cv2.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "PIL.*"
ignore_missing_imports = true

# Black configuration
[tool.black]
line-length = 88
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["tilapia_freshness"]

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
